# samples/lambda-blueprint/src/apidocs.yml
openapi: 3.0.0
info:
  title: User Profile API
  description: API to manage user profiles
  version: 1.0.0
servers:
  - url: http://localhost:8000
paths:
  /profiles:
    post:
      summary: Create user profile
      description: Creates a new user profile
      operationId: createUserProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserProfileInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /profiles/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        description: ID of the user profile
        schema:
          type: string
          format: uuid
    get:
      summary: Get user profile by ID
      description: Returns a single user profile based on the provided ID
      operationId: getUserProfileById
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: User profile not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    delete:
      summary: Delete user profile by ID
      description: Deletes a single user profile based on the provided ID
      operationId: deleteUserProfileById
      responses:
        '204':
          description: User profile deleted
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: User profile not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
components:
  schemas:
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: array
          items: {}
      required:
        - type
        - title
        - status
    CreateUserProfileInput:
      allOf:
        - $ref: '#/components/schemas/UserProfileData'
    UserProfileData:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
      required:
        - firstName
        - lastName
        - email
    UserProfile:
      allOf:
        - $ref: '#/components/schemas/UserProfileData'
        - type: object
          properties:
            id:
              type: string
              format: uuid
          required:
            - id